extract function for parsing JS object
