+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "./exampleCode";
 export * from "./formatting";
+export * from "./parseObj";
 export * from "./transformers";
 export * from "./usePrevious";
