+++ src/utils/parseObj.ts
@@ -1,0 +1,23 @@
+/**
+ * Loosely parses the specified string as a JS object.
+ */
+export function parseObj(objString) {
+  const rules = {};
+
+  objString.split("\n").forEach(line => {
+    line = line.replace(/,$/, ""); // remove trailing comma
+
+    // Split each line into a key and a value
+    // Everything before the first : is the key and the reset is the value
+    const segments = line.split(":");
+    if (segments.length < 2) return; // skip this line
+    const key = segments[0].trim();
+    const value = segments
+      .slice(1)
+      .join(":")
+      .trim();
+    rules[key] = value;
+  });
+
+  return rules;
+}
